permissoes

+++ app/Http/Controllers/Api/NoticiaController.php
@@ -15,6 +15,11 @@ class NoticiaController extends Controller
 
     public function store(Request $request)
     {
+        $user = auth('sanctum')->user();
+
+        if (!$user->can('create', Noticia::class)) {
+            return response()->json('Não Autorizado', 401);
+        }
         $noticia = Noticia::create($request->all());
         return response()->json($noticia, 201);
     }
@@ -23,7 +28,7 @@ class NoticiaController extends Controller
     {
         $user = auth('sanctum')->user();
         if (!$user->can('view', $noticia)) {
-            return response()->json('Nao Autorizado', 401);
+            return response()->json('Não Autorizado', 401);
         }
 
         return response()->json($noticia, 200);
@@ -31,6 +36,11 @@ class NoticiaController extends Controller
 
     public function update(Noticia $noticia, Request $request)
     {
+        $user = auth('sanctum')->user();
+
+        if (!$user->can('update', $noticia)) {
+            return response()->json('Não Autorizado', 401);
+        }
         $noticia->titulo = $request->titulo;
         $noticia->descricao = $request->descricao;
         $noticia->user_id = $request->user_id;
@@ -41,6 +51,11 @@ class NoticiaController extends Controller
 
     public function destroy(Noticia $noticia)
     {
+        $user = auth('sanctum')->user();
+
+        if (!$user->can('delete', $noticia)) {
+            return response()->json('Nao Autorizado', 401);
+        }
         Noticia::destroy($noticia->id);
 
         return response()->noContent();
